+++ R/argo.R
@@ -1687,8 +1687,6 @@ setMethod(f="plot",
           {
               if (!inherits(x, "argo"))
                   stop("method is only for objects of class '", "argo", "'")
-              if ("adorn" %in% names(list(...)))
-                  warning("In plot,argo-method() : the 'adorn' argument was removed in November 2017", call.=FALSE)
               oceDebug(debug, "plot.argo(x, which=c(", paste(which, collapse=","), "),",
                       argShow(mgp),
                       argShow(mar),